refactor: Move the `database query` command to `query`
